backend/requirements.txt
amqp==5.1.1
asgiref==3.6.0
attrs==22.2.0
autobahn==23.1.1
Automat==22.10.0
billiard==3.6.4.0
black==23.1.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.1.0
cryptography==39.0.0
daphne==4.0.0
Django==4.1.6
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-polymorphic==3.1.0
django-rest-polymorphic==0.1.10
djangorestframework==3.14.0
drf-yasg==1.21.5
exceptiongroup==1.1.0
execnet==1.9.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
platformdirs==3.0.0
pluggy==1.0.0
prompt-toolkit==3.0.36
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.0.0
pytest==7.2.1
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-sugar==0.9.6
pytest-xdist==3.2.0
pytz==2022.7.1
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
termcolor==2.2.0
tomli==2.0.1
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==23.1.1
typing_extensions==4.4.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
zope.interface==5.5.2
